// QuickApp/ClientApp/src/app/models/insurance-company.ts
export class InsuranceCompany {
    name: string;
    logo: string;
    logoDescription: string;
    address: string;
    city: string;
    postalCode: string;
    country: string;
    phoneNumber: string;
    faxNumber: string;
    email: string;
    abstract: string;   
    description: string;
    capitaleSociale: string;
    descriptionList: [string];
}